Add files via upload
n a two-dimensional space, this hyperplane is simply a line that divides the data points into two groups. The algorithm identifies the data points that are closest to the boundary, known as support vectors, and uses them to define the position of the hyperplane. By maximizing the margin between classes, SVM aims to reduce classification errors and improve the model’s generalization to unseen data.